Consolidate example, lint prettier again
