+++ package.json
@@ -7,7 +7,9 @@
     "build": "ng build",
     "test": "ng test",
     "lint": "ng lint",
-    "prettier": "prettier \"src/**/**.{ts,tsx,js,jsx,json,css,scss,html}\" --write"
+    "prettier": "prettier \"src/**/**.{ts,tsx,js,jsx,json,css,scss,html}\" --write",
+      "lint": "eslint \"src/**/**.ts\"",
+      "lint:fix": "eslint \"src/**/**.ts\" --fix"
   },
   "prettier": "@pxblue/prettier-config",
   "private": true,
@@ -39,11 +41,16 @@
     "@angular/cli": "~9.0.7",
     "@angular/compiler-cli": "~9.1.6",
     "@angular/language-service": "~9.1.6",
+    "@pxblue/eslint-config": "^1.1.5",
     "@pxblue/prettier-config": "^1.0.2",
     "@types/jasmine": "~3.5.0",
     "@types/jasminewd2": "~2.0.3",
     "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "^2.33.0",
+    "@typescript-eslint/parser": "^2.33.0",
     "codelyzer": "^5.1.2",
+    "eslint": "^7.0.0",
+    "eslint-config-prettier": "^6.11.0",
     "jasmine-core": "~3.5.0",
     "jasmine-spec-reporter": "~4.2.1",
     "karma": "~4.3.0",
