+++ src/app/app.component.html
@@ -16,15 +16,90 @@
             <mat-divider></mat-divider>
             <div class="navWrapper">
                 <section fxLayout="column">
-                    <a mat-list-item *ngFor="let item of menuItems; let i = index" (click)="toggleMenu()">
-                        <mat-icon matListIcon class="icon">{{ menuIcons[i] }}</mat-icon>
-                        <span>{{ 'MENU_ITEMS.' + item | translate }}</span>
+                    <a mat-list-item (click)="toggleMenu()">
+                        <mat-icon matListIcon class="icon">home</mat-icon>
+                        <span>{{ 'MENU_ITEMS.HOME' | translate }}</span>
+                    </a>
+                    <a mat-list-item (click)="toggleMenu()">
+                        <mat-icon matListIcon class="icon">folder</mat-icon>
+                        <span>{{ 'MENU_ITEMS.PROJECTS' | translate }}</span>
+                    </a>
+                    <a mat-list-item (click)="toggleMenu()">
+                        <mat-icon matListIcon class="icon">report</mat-icon>
+                        <span>{{ 'MENU_ITEMS.WARNINGS' | translate }}</span>
+                    </a>
+                    <a mat-list-item (click)="toggleMenu()">
+                        <mat-icon matListIcon class="icon">settings</mat-icon>
+                        <span>{{ 'MENU_ITEMS.SETTINGS' | translate }}</span>
+                    </a>
+                    <a mat-list-item (click)="toggleMenu()">
+                        <mat-icon matListIcon class="icon">help</mat-icon>
+                        <span>{{ 'MENU_ITEMS.HELP' | translate }}</span>
                     </a>
                 </section>
             </div>
         </mat-nav-list>
     </mat-sidenav>
     <mat-sidenav-content class="nav-content">
-        <list-component (menuClicked)="toggleMenu()"></list-component>
+        <mat-toolbar color="primary">
+            <button
+                type="button"
+                aria-label="Toggle sidenav"
+                mat-icon-button
+                (click)="toggleMenu()"
+                style="margin-left: -4px; margin-right: 12px;"
+            >
+                <mat-icon aria-label="Side nav toggle icon">menu</mat-icon>
+            </button>
+            <div>{{ 'I18N' | translate }}</div>
+            <div fxFlex style="margin: auto;"></div>
+        </mat-toolbar>
+
+        <mat-toolbar class="secondary-toolbar">
+            <mat-form-field>
+                <mat-select
+                    #langeSelect
+                    (ngModelChange)="translate.use(langeSelect.value)"
+                    [(ngModel)]="selectedLanguage"
+                >
+                    <mat-option *ngFor="let lang of enabledLocales; let i = index" [value]="enabledLocales[i]">
+                        {{ 'LANGUAGES.' + lang | translate }}
+                    </mat-option>
+                </mat-select>
+            </mat-form-field>
+            <button mat-raised-button color="primary" style="margin: 20px;">
+                <mat-icon>shopping_cart</mat-icon> {{ 'ADD_TO_CART' | translate }}
+            </button>
+        </mat-toolbar>
+
+        <mat-nav-list>
+            <mat-list-item *ngFor="let fruit of fruits; let i = index" (click)="toggleFruit(fruit)">
+                <mat-checkbox
+                    mat-list-icon
+                    [checked]="selectedFruits.has(fruit)"
+                    (change)="toggleFruit(fruit)"
+                    color="primary"
+                ></mat-checkbox>
+                <div mat-line>
+                    {{ 'FRUITS.' + fruit | translate }}
+                </div>
+                <div mat-line class="list">
+                    {{ 'MORE_INFO' | translate }}
+                </div>
+                <mat-icon class="icon">navigate_next</mat-icon>
+            </mat-list-item>
+        </mat-nav-list>
+
+        <footer [class.active]="selectedFruits.size > 0" class="mat-simple-snackbar" fxLayout="row">
+            <p style="">{{ selectedFruits.size }} {{ 'ITEMS' | translate }}</p>
+            <div class="icon">
+                <button mat-icon-button (click)="cancelItems()">
+                    <mat-icon>cancel</mat-icon>
+                </button>
+                <button mat-icon-button>
+                    <mat-icon>more_vert</mat-icon>
+                </button>
+            </div>
+        </footer>
     </mat-sidenav-content>
 </mat-sidenav-container>
