+++ src/app/app.component.ts
@@ -1,5 +1,5 @@
 import { Component } from '@angular/core';
-import { TranslateService } from '@ngx-translate/core';
+import { TranslateService, TranslationChangeEvent } from '@ngx-translate/core';
 import { SampleTranslation } from './i18n/sample-translation';
 import { english } from './i18n/english';
 
@@ -10,16 +10,38 @@ import { english } from './i18n/english';
 })
 export class AppComponent {
     open = false;
-    menuIcons = ['home', 'folder', 'report', 'settings', 'help'];
-    menuItems = Object.keys(english.MENU_ITEMS);
+    enabledLocales: Array<keyof SampleTranslation['LANGUAGES']> = ['EN', 'ES', 'FR', 'DE', 'PT', 'ZH', 'AR'];
+    selectedLanguage: string;
+    selectedFruits = new Set<string>();
+    fruits = Object.keys(english.FRUITS);
 
     constructor(public translate: TranslateService) {
-        const enabledLocales: Array<keyof SampleTranslation['LANGUAGES']> = ['EN', 'ES', 'FR', 'DE', 'PT', 'ZH', 'AR'];
-        translate.addLangs(enabledLocales);
+        translate.addLangs(this.enabledLocales);
         translate.setDefaultLang('EN');
+        this.selectedLanguage = this.enabledLocales[0];
+        this.listenForLanguageChanges();
     }
 
-    toggleMenu() {
+    toggleMenu(): void {
         this.open = !this.open;
+    }
+
+    toggleFruit(fruit: string): void {
+        this.selectedFruits.has(fruit) ? this.selectedFruits.delete(fruit) : this.selectedFruits.add(fruit);
+    }
+
+    isRTL(lang: string): boolean {
+        return lang === 'AR';
+    }
+
+    cancelItems(): void {
+        this.selectedFruits.clear();
+    }
+
+    listenForLanguageChanges(): void {
+        this.translate.onLangChange.subscribe((event: TranslationChangeEvent) => {
+            const body = document.querySelector('body');
+            body.setAttribute('dir', this.isRTL(event.lang) ? 'rtl' : 'ltr');
+        });
     }
 }
