+++ src/app/app.module.ts
@@ -4,10 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { LayoutModule } from '@angular/cdk/layout';
 import { FlexLayoutModule } from '@angular/flex-layout';
-
 import { AppComponent } from './app.component';
-
-import { HttpClient, HttpClientModule } from '@angular/common/http';
 import { MatBottomSheetModule } from '@angular/material/bottom-sheet';
 import { MatButtonModule } from '@angular/material/button';
 import { MatCheckboxModule } from '@angular/material/checkbox';
@@ -20,10 +17,9 @@ import { MatSidenavModule } from '@angular/material/sidenav';
 import { MatToolbarModule } from '@angular/material/toolbar';
 import { TranslateModule, TranslateLoader } from '@ngx-translate/core';
 import { LanguageLoader } from './language-loader';
-import { ListComponent } from './list/list.component';
 
 @NgModule({
-    declarations: [AppComponent, ListComponent],
+    declarations: [AppComponent],
     imports: [
         BrowserAnimationsModule,
         BrowserModule,
@@ -40,11 +36,8 @@ import { ListComponent } from './list/list.component';
         MatSelectModule,
         MatSidenavModule,
         MatToolbarModule,
-        HttpClientModule,
-
         TranslateModule.forRoot({
             loader: { provide: TranslateLoader, useClass: LanguageLoader },
-
             defaultLanguage: 'EN',
         }),
     ],
